Example does not work; brew install too invasive
Thank you for publishing this module. While it is a fine start, I could not get
it to work. This pull request contains my files at the point where I gave up
trying to make it work with Cognito.

This pull request just illustrates the issues I found; it is not intended to be
merged as-is.

Here are the issues I found:
- The example is incomplete. It does not specify required variables.
- I am on Ubuntu 18.04 and I do not care to install `brew`
- I think it is inappropriate for Terraform to install packages and modify user
  configuration
- `cloudfront_auth_brach` is misspelled
- I am unsure of my assumptions about the nodejs version--I did not get that far
- In `variables.tf` I commented on variables that needed more explanation