Removing homebrew and nvm.sh
Removing the local usage of homebrew and sourcing nvm.sh as this is
outside of the scope of Terraform.

When using the module the environment should be setup to have nvm and
enable the use of it. This is enough to allow building in other
environments and build systems like Jenkins or Atlantis.