Remove homebrew and trust nvm is available.
Removing homebrew as that is platform specific and trusting nvm is
available to set the nodejs version.

NVM is being used with the -s flag during initial install to ensure it
works on Alpine Linux which has to have custom nodejs versions compiled
due to the libc difference.

This is going to cause the module to take longer to run.